requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch
aiogram
llama-cpp-python==0.1.83
